requirements.txt
flask
numpy==1.26.4
Pillow
colour-science==0.4.4
psutil
scipy
piexif